Fixed code to summon enemies mid-battle and added more placeholder sprites
The code to summon enemies during battle now relies on a list of positions that can either be taken up, or not taken up. It will choose among the list to render an untaken position, rather than choose a position based on how many enemies are remaining. Also added a unique summinion enemy that is summoned by the summoner, which heals the entire enemy party