Lots of bug fixing
Fixed an offset bug where the item tooltip did not show up. Fixed a bug where moving the enemy sprite pivots overlapped on the status object and did not allow for the status tooltip to show. Fixed a bug where renaming the Tree? to Treevil prevented it from gaining Growth